% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:run_app.R
\name{run_kmeans_app}
\alias{run_kmeans_app}
\title{Launch K-means Clustering Shiny App}
\usage{
run_kmeans_app()
}
\value{
A Shiny app object
}
\description{
This function launches the interactive Shiny application for K-means clustering.
}
\examples{
\dontrun{
run_kmeans_app()
}
}
